% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{add_class}
\alias{add_class}
\title{Add a class membership to any object}
\usage{
add_class(x, name)
}
\arguments{
\item{x}{stars object}

\item{name}{chr, the class name(s) to add to the object}
}
\value{
the input, `x`, which class augmentation
}
\description{
This is handy when slicing/filtering stars objects which may drop 
any extra class label you hmay have added earlier
}
