% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{build_database}
\alias{build_database}
\title{Build a database by mining for files with a particualr extension}
\usage{
build_database(
  path = copernicus::copernicus_path("mthw", "chfc"),
  save_db = FALSE,
  filename = "database",
  ext = ".tif"
)
}
\arguments{
\item{path}{chr, the path description}

\item{save_db}{logical, if true then save the database under `filename`}

\item{filename}{chr the database filename to save under}

\item{ext}{chr, the file extension (with dot) that defines a record}

\item{a}{tabular database}
}
\description{
Build a database by mining for files with a particualr extension
}
