% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{colors_mwe}
\alias{colors_mwe}
\title{Retrieve a standard mwe (event) color scheme}
\usage{
colors_mwe(
  breaks = c(-1, -0.9, 0, 0.9, 1),
  values = scales::rescale(breaks),
  labels = c("", "< q10", "q10 to q90", "> q90", ""),
  colors = c("#ffffff", scales::muted("blue"), "#ffffff", scales::muted("red"),
    "#ffffff")
)
}
\value{
ggplot2::scale_fill_stepsn object
}
\description{
Retrieve a standard mwe (event) color scheme
}
