% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{compose_filename}
\alias{compose_filename}
\title{Compose one or more filenames from a tabular database}
\usage{
compose_filename(
  x = decompose_filename(),
  path = andreas::copernicus_path("mthw"),
  ext = ".tif"
)
}
\arguments{
\item{x}{tabular database}

\item{path}{chr, the path to where the data resides}

\item{ext}{chr, the extension to apply (with dot)}
}
\value{
filenames
}
\description{
Compose one or more filenames from a tabular database
}
