% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename}
\alias{decompose_filename}
\title{Decompose one or more filenames into a simple tabular database}
\usage{
decompose_filename(
  x = c("/mnt/s1/projects/ecocast/coredata/copernicus/mthw/chfc/temp/bot/q10/011.tif",
    "/mnt/s1/projects/ecocast/coredata/copernicus/mthw/chfc/temp/sur/q10/353.tif",
    "/mnt/s1/projects/ecocast/coredata/copernicus/mthw/chfc/sal/sur/q90/088.tif",
    "/mnt/s1/projects/ecocast/coredata/copernicus/mthw/chfc/temp/bot/q10/363.tif",
    "/mnt/s1/projects/ecocast/coredata/copernicus/mthw/chfc/temp/sur/q90/164.tif"),
  ext = andreas::get_extension(x)[1]
)
}
\arguments{
\item{x}{chr, one or more filenames (with complete paths)}

\item{ext}{chr, the file extension to strip including the dot}
}
\value{
tabular database
}
\description{
Decompose one or more filenames into a simple tabular database
}
