% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{encode_wave}
\alias{encode_wave}
\title{Given marine wave event data, compute duration of waves}
\usage{
encode_wave(x = generate_mtw(), event_window = 5, gap_width = 2, mask = TRUE)
}
\arguments{
\item{x}{stars object that includes 'mwe'}

\item{event_window}{num, the number of days needed to define an event}

\item{gap_width}{num, when gaps of this or fewer days occur between events, 
try to merge the events}

\item{mask}{logical, if TRUE, then mask the result so only events of 
event_window days are returned.  Shorter events are set to 0.}
}
\value{
stars object with records of thermal waves duration (hence "mwd" attribute name).
}
\description{
Given marine wave event data, compute duration of waves
}
