% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{generate_wave}
\alias{generate_wave}
\title{Generate the MTH wave data}
\usage{
generate_wave(
  db = dplyr::filter(read_database(), name == "temp", depth == "sur"),
  dates = c(Sys.Date() - 5, Sys.Date() + 10),
  region = dplyr::pull(db, dplyr::all_of("region"))[1],
  cDB = andreas::read_database(andreas::copernicus_path(region), multi = TRUE)
)
}
\arguments{
\item{db}{tabular database for a region}

\item{dates}{Date, two element start and stop}

\item{region}{}
}
\value{
stars object with sst (thetao), q10, q90 and mwe (marine weather event)
   heating events are positive, cooling events are negative
}
\description{
Generate the MTH wave data
}
