% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{has_time}
\alias{has_time}
\title{Check if a stars object has a time dimension}
\usage{
has_time(x, names = c("time", "date", "doy"))
}
\arguments{
\item{x}{stars object}

\item{names}{chr, possible time related dimension names}
}
\value{
logical
}
\description{
Check if a stars object has a time dimension
}
