% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{mround}
\alias{mround}
\title{Round numbers to nearest base value}
\usage{
mround(x, base)
}
\arguments{
\item{x}{numeric numbers to round}

\item{base}{number, round to the nearest `base`}
}
\description{
Round numbers to nearest base value
}
\seealso{
https://stat.ethz.ch/pipermail/r-help/2009-August/401258.html
}
