% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthw-package.R
\docType{package}
\name{mthw}
\alias{mthw-package}
\alias{mthw}
\title{mthw package}
\description{
Detecting and forecasting Marine Thermohaline Waves using Copernicus data in R programming language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/mthw}
}

}
\author{
\strong{Maintainer}: Tupper Ben \email{btupper@bigelow.org}

Other contributors:
\itemize{
  \item Oliveira Kyle \email{kyle.oliveira@maine.edu} [contributor]
}

}
