% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mthw_filename}
\alias{mthw_filename}
\title{Generate a mthw-centric filename}
\usage{
mthw_filename(
  region = "chfc",
  variable = "temp",
  depth = "sur",
  extension = c(".rds", ".png", ".pdf", ".jpg")[1],
  path = c("/mnt/ecocast/corecode/R/mthw/inst", system.file(package = "mthw"))[1]
)
}
\arguments{
\item{region}{chr the region code}

\item{variable}{chr a single variable to read}

\item{depth}{chr a single depth to read}

\item{extension}{chr, the file extension (with dot)}

\item{path}{chr, the file path}
}
\description{
Generate a mthw-centric filename
}
