% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{mwd_breaks}
\alias{mwd_breaks}
\title{Generate breaks for the display of encoded wave data}
\usage{
mwd_breaks(x = encode_wave(), base = 5)
}
\arguments{
\item{x}{stars object, encoded wave event durations}

\item{base}{number, round the breaks to the nearest `base`}
}
\value{
number, sequence of break points
}
\description{
Generate breaks for the display of encoded wave data
}
