% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_mwd}
\alias{plot_mwd}
\title{Plot a set of 'mwd' rasters}
\usage{
plot_mwd(x = encode_wave(), y = NULL, show.legend = TRUE, title = "", ...)
}
\arguments{
\item{x}{stars 'mwd' raster ala `encode_wave`}

\item{y}{ignored}

\item{show.legend}{logical, shown if TRUE}

\item{...}{arguments for facet_wrap}
}
\value{
ggplot2 object
}
\description{
Plot a set of 'mwd' rasters
}
