% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_mwd_list}
\alias{plot_mwd_list}
\title{Plot a paired wave event or duration for a single day}
\usage{
plot_mwd_list(x, title = "Marine Thermohaline Waves")
}
\arguments{
\item{x}{named list of stars stars objects with "mwd"}

\item{title}{str, the title for the plot}
}
\description{
Plot a paired wave event or duration for a single day
}
