% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_mwd_paired}
\alias{plot_mwd_paired}
\title{Plot a paired wave even or duration for a single day}
\usage{
plot_mwd_paired(temp, sal, title = "Marine Thermohaline Waves")
}
\arguments{
\item{temp}{stars object with "mwd" - perhaps temp}

\item{sal}{stars object with "mwd" - say for sal}
}
\description{
Plot a paired wave even or duration for a single day
}
