% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_mwe}
\alias{plot_mwe}
\title{Plot a set of 'mwe' rasters}
\usage{
plot_mwe(
  x = generate_wave(),
  y = NULL,
  show.legend = has_time(x),
  title = "",
  ...
)
}
\arguments{
\item{x}{stars 'mwe' raster ala `generate_wave`}

\item{y}{ignored}

\item{show.legend}{logical, shown if TRUE}

\item{ncol}{NULL or numeric number of columns}
}
\value{
ggplot2 object
}
\description{
Plot a set of 'mwe' rasters
}
