% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_database}
\alias{read_database}
\title{Read a database files}
\usage{
read_database(
  path = andreas::copernicus_path("mthw", "chfc"),
  filename = "database"
)
}
\arguments{
\item{path}{chr the path to the database}
}
\description{
Read a database files
}
