% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_mthw}
\alias{read_mthw}
\title{Read a database of MTHW data into a stars object}
\usage{
read_mthw(
  db = dplyr::filter(read_database(), doy \%in\% c("001", "002", "003"), name \%in\%
    c("temp", "sal")),
  path = andreas::copernicus_path("mthw"),
  form = c("stars", "table")[1]
)
}
\arguments{
\item{db}{a tabular database that must be symmetric about name, depth and percentile}

\item{path}{chr, the path to the database}
}
\value{
stars object of [name_depth_percentile, x, y, doy] or the input table augmented
  with a 'data' list column of individual stars objects
}
\description{
Read a database of MTHW data into a stars object
}
