% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave.R
\name{slice_date}
\alias{slice_date}
\title{A slicer that retains class}
\usage{
slice_date(x, date)
}
\arguments{
\item{x}{stars object}

\item{date}{numeric or Date identifying the slabs of time to slice}

\item{the}{input sliced and with its original class label}
}
\description{
A slicer that retains class
}
