% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_raster}
\alias{write_raster}
\alias{read_raster}
\title{Read and write mthw raster data}
\usage{
write_raster(x, filename)

read_raster(filename = mthw_filename(path = system.file(package = "mthw")))
}
\arguments{
\item{x}{stars object}

\item{filename}{chr then name of the file}
}
\value{
for `write_raster` and `read_raster` a stars object, while 
  `get_filename` returns a characater file path
}
\description{
Read and write mthw raster data
}
