% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{colors_mwd}
\alias{colors_mwd}
\title{Retrieve a standard mwd (duration) color scheme}
\usage{
colors_mwd(
  colors = rev(c("#b2182b", "#ef8a62", "#ffffff", "#ffffff", "#67a9cf", "#2166ac")),
  breaks = NULL,
  labels = NULL
)
}
\arguments{
\item{colors}{chr, color palette}

\item{labels}{chr, labels (length must be one more than colors)}
}
\value{
ggplot2::scale_fill_binned object
}
\description{
Retrieve a standard mwd (duration) color scheme
}
